% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPD.fit.R
\name{GPD.fit}
\alias{GPD.fit}
\title{Fit the non-stationary GPD model to skew surges}
\usage{
GPD.fit(
  data,
  q = 0.95,
  optim.method = "BFGS",
  init.par = c(0.2, 0, 100, 0.1, 0.1)
)
}
\arguments{
\item{data}{A data frame of skew surge observations (named \code{skews}), along with \code{month}, \code{day} and \code{maxTide} observations.}

\item{q}{The monthly quantile of skew surges to obtain exceedances and must be between 0 and 1. The default is 0.95 but this should be investigated using standard threshold selection techniques.}

\item{optim.method}{The method to be used for optimisation of the log-likelihood function for the GPD. The default is \code{'BFGS'} but see \code{\link[stats]{optim}} for details and other options.}

\item{init.par}{The initial values for parameters to be optimised over in \code{\link[stats]{optim}}. The default are \code{c(0.2,0,100,0.1,0.1)} but these can be altered.}
}
\value{
Parameters for the GPD model fit: 4 parameters characterising the scale parameter and 1 shape parameter.
}
\description{
The function fits a non-stationary generalised Pareto distribution (GPD) to skew surge exceedances of a monthly threshold. This has a non-stationary scale parameter to represent seasonal variability (via a harmonic) and skew surge-peak tide dependence (via a linear trend).
}
\details{
\loadmathjax{} This function uses maximum likelihood estimation to fit a non-stationary GPD to skew surge exceedances of a monthly threshold (determined by the quantile \mjeqn{q}{q}).
Non stationarity is introduced to the scale parameter of the GPD using a harmonic to capture seasonal variations and a linear trend for tide, to capture skew surge-peak tide dependence. 
This parametrisation is given by \mjeqn{\sigma_{d,t}=\alpha+\beta\sin[(2\pi/f)(d-\phi)]+\gamma x}{sigma_{d,t}=alpha + beta*sin[((2pi)/f)*(d-phi)]+gamma*x} for \mjeqn{\alpha,\beta,\phi,\gamma}{alpha, beta, phi, gamma} parameters to be estimated, \mjeqn{x}{x} the peak tide observation, \mjeqn{d\in[1:365]}{d} the day in year covariate and \mjeqn{f=365}{f=365} the periodicity of the harmonic.

The GPD log likelihood is given by the function \code{GPD_ll} within \code{GPD.fit} and then optimised using \code{\link[stats]{optim}} to obtain parameter estimates for \mjeqn{\sigma}{sigma} (i.e., \mjeqn{\alpha,\beta,\phi,\gamma}{alpha, beta, phi, gamma}) and \mjeqn{\xi}{xi} the GPD shape parameter.

A Normal prior distribution is incorporated onto the log likelihood to reduce uncertainty associated with shape parameter estimation. 
This prior information is based on spatial information about shape parameter estimates across the UK - see the Coastal Flood Boundary report (2018).

Further details can be found in Section 4 of \insertCite{DArcy2023;textual}{ESLestimation}.
}
\examples{

library(ESLestimation)

# Load in data
# This example uses Lowestoft data available within the package 

data(Lowestoft)

# Check the data has the required variables that are correctly named (skews, month, day, maxTide)
head(Lowestoft)

# Fit the non-stationary GPD model to exceedances of the 0.95 monthly quantile.
# In this example we use the standard arguments for q, optim.method and init.par
( gpd_par <- GPD.fit(Lowestoft) )

}
\references{
\insertRef{DArcy2023}{ESLestimation}
}
